chore: Refactor My Collection upload flow (#11539)
* chore: Refactor My Collection upload flow

* fix tests

* introduce contesxt

* change next button
