+++ .secrets.baseline
@@ -127,13 +127,13 @@
         "line_number": 22
       }
     ],
-    "src/Apps/MyCollection/Routes/EditArtwork/__tests__/MyCollectionArtworkForm.jest.tsx": [
+    "src/Apps/MyCollection/Routes/EditArtwork/__tests__/MyCollectionEditArtwork.jest.tsx": [
       {
         "type": "Base64 High Entropy String",
-        "filename": "src/Apps/MyCollection/Routes/EditArtwork/__tests__/MyCollectionArtworkForm.jest.tsx",
+        "filename": "src/Apps/MyCollection/Routes/EditArtwork/__tests__/MyCollectionEditArtwork.jest.tsx",
         "hashed_secret": "dc312ec7ed341804d6911ecdcf18eea1858f6091",
         "is_verified": false,
-        "line_number": 538
+        "line_number": 484
       }
     ],
     "src/Apps/ViewingRoom/Routes/Statement/__tests__/ViewingRoomStatementRoute.jest.tsx": [
@@ -1232,37 +1232,7 @@
         "is_verified": false,
         "line_number": 182
       }
-    ],
-    "src/Components/NavBar/Menus/__tests__/NavBarNotifications.jest.tsx": [
-      {
-        "type": "Base64 High Entropy String",
-        "filename": "src/Components/NavBar/Menus/__tests__/NavBarNotifications.jest.tsx",
-        "hashed_secret": "25e8aedd607dba94e94c57ca562b4bba38d00ca1",
-        "is_verified": false,
-        "line_number": 63
-      },
-      {
-        "type": "Base64 High Entropy String",
-        "filename": "src/Components/NavBar/Menus/__tests__/NavBarNotifications.jest.tsx",
-        "hashed_secret": "0ee07b90c46b9bbec6fa7be2947286ec7cf81a95",
-        "is_verified": false,
-        "line_number": 81
-      },
-      {
-        "type": "Base64 High Entropy String",
-        "filename": "src/Components/NavBar/Menus/__tests__/NavBarNotifications.jest.tsx",
-        "hashed_secret": "071bef2b491d817c0922c5c5a8f6d41fd4702d76",
-        "is_verified": false,
-        "line_number": 99
-      },
-      {
-        "type": "Base64 High Entropy String",
-        "filename": "src/Components/NavBar/Menus/__tests__/NavBarNotifications.jest.tsx",
-        "hashed_secret": "4a31efa2939f52a6ff9afa32986ac8fbda578d6c",
-        "is_verified": false,
-        "line_number": 117
-      }
     ]
   },
-  "generated_at": "2022-12-14T12:47:10Z"
+  "generated_at": "2022-12-14T13:43:11Z"
 }
