+++ src/Apps/MyCollection/Routes/EditArtwork/Components/ConfirmationModalDelete.tsx
