+++ src/Apps/MyCollection/Routes/EditArtwork/Utils/useCreateOrUpdateArtwork.ts
@@ -1,7 +1,12 @@
-import { MyCollectionCreateArtworkInput } from "__generated__/useCreateArtworkMutation.graphql"
+import { ArtworkModel } from "Apps/MyCollection/Routes/EditArtwork/Utils/artworkModel"
+import { MyCollectionEditArtwork_artwork$data } from "__generated__/MyCollectionEditArtwork_artwork.graphql"
+import {
+  ArtworkAttributionClassType,
+  MyCollectionCreateArtworkInput,
+} from "__generated__/useCreateArtworkMutation.graphql"
 import { MyCollectionUpdateArtworkInput } from "__generated__/useUpdateArtworkMutation.graphql"
-import { useCreateArtwork } from "../Mutations/useCreateArtwork"
-import { useUpdateArtwork } from "../Mutations/useUpdateArtwork"
+import { useCreateArtwork } from "Apps/MyCollection/Routes/EditArtwork/Mutations/useCreateArtwork"
+import { useUpdateArtwork } from "Apps/MyCollection/Routes/EditArtwork/Mutations/useUpdateArtwork"
 
 export type ArtworkInput =
   | MyCollectionCreateArtworkInput
@@ -11,13 +16,18 @@ export const useCreateOrUpdateArtwork = () => {
   const { submitMutation: createArtwork } = useCreateArtwork()
   const { submitMutation: updateArtwork } = useUpdateArtwork()
 
-  const createOrUpdateArtwork = async (artwork: ArtworkInput) => {
+  const createOrUpdateArtwork = async (
+    values: ArtworkModel,
+    artwork?: MyCollectionEditArtwork_artwork$data
+  ) => {
+    const artworkInputValues = formValuesToMutationInput(values, artwork)
+
     let artworkId: string | undefined
 
-    if ((artwork as MyCollectionUpdateArtworkInput).artworkId) {
+    if ((artworkInputValues as MyCollectionUpdateArtworkInput).artworkId) {
       const res = await updateArtwork({
         variables: {
-          input: artwork as MyCollectionUpdateArtworkInput,
+          input: artworkInputValues as MyCollectionUpdateArtworkInput,
         },
         rejectIf: res => {
           return res.myCollectionUpdateArtwork?.artworkOrError?.mutationError
@@ -29,7 +39,7 @@ export const useCreateOrUpdateArtwork = () => {
     } else {
       const res = await createArtwork({
         variables: {
-          input: artwork as MyCollectionCreateArtworkInput,
+          input: artworkInputValues as MyCollectionCreateArtworkInput,
         },
       })
 
@@ -42,4 +52,44 @@ export const useCreateOrUpdateArtwork = () => {
   }
 
   return { createOrUpdateArtwork }
+}
+
+const formValuesToMutationInput = (
+  values: ArtworkModel,
+  artwork?: MyCollectionEditArtwork_artwork$data
+): ArtworkInput => {
+  // Set edition values for attribution class
+
+  if (values.attributionClass !== "LIMITED_EDITION") {
+    values.editionNumber = ""
+    values.editionSize = ""
+  }
+
+  const externalImageUrls = values.newPhotos.flatMap(photo => photo.url || null)
+
+  return {
+    artworkId: artwork?.internalID,
+    artistIds: [values.artistId],
+    category: values.category,
+    date: values.date,
+    title: values.title,
+    medium: values.medium,
+    attributionClass: values.attributionClass
+      ?.replace(" ", "_")
+      ?.toUpperCase() as ArtworkAttributionClassType,
+    editionNumber: String(values.editionNumber),
+    editionSize: String(values.editionSize),
+    height: String(values.height),
+    width: String(values.width),
+    depth: String(values.depth),
+    metric: values.metric,
+    externalImageUrls,
+    pricePaidCents:
+      !values.pricePaidDollars || isNaN(Number(values.pricePaidDollars))
+        ? undefined
+        : Number(values.pricePaidDollars) * 100,
+    pricePaidCurrency: values.pricePaidCurrency,
+    provenance: values.provenance,
+    artworkLocation: values.artworkLocation,
+  }
 }
